# apps/worker/openapi-restful.yaml
openapi: 3.0.3
info:
  title: Sutradhar RESTful Actions API
  description: |
    RESTful API endpoints for GitHub, Slack, and Google Calendar operations.
    
    All endpoints follow REST principles with resource-based URLs, proper HTTP methods,
    and consistent response formats.
  version: 1.0.0
servers:
  - url: http://localhost:2198/api/v1
    description: Development server
...
paths:
...
  /slack/channels:
    get:
      tags: [Slack]
      summary: List channels
      parameters:
        - name: types
          in: query
          schema:
            type: string
        - name: excludeArchived
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of channels
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: object
                    properties:
                      channels:
                        type: array
                        items:
                          type: object
                      count:
                        type: integer
  /slack/channels/{channelId}:
    get:
      tags: [Slack]
      summary: Get channel information
      parameters:
        - name: channelId
          in: path
          required: true
      responses:
        '200':
          description: Channel information
  /slack/channels/{channelId}/messages:
    get:
      tags: [Slack]
      summary: List messages in channel
      parameters:
        - name: channelId
          in: path
          required: true
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: cursor
          in: query
          schema:
            type: string
        - name: oldest
          in: query
          schema:
            type: string
        - name: latest
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: object
                    properties:
                      messages:
                        type: array
                        items:
                          type: object
                      count:
                        type: integer

    post:
      tags: [Slack]
      summary: Send message to channel
      parameters:
        - name: channelId
          in: path
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  minLength: 1
                threadTs:
                  type: string
                blocks:
                  type: array
                sessionId:
                  type: string
      responses:
        '201':
          description: Message sent
  /slack/channels/{channelId}/messages/{ts}:
    get:
      tags: [Slack]
      summary: Get message
      parameters:
        - name: channelId
          in: path
          required: true
        - name: ts
          in: path
          required: true
      responses:
        '200':
          description: Message details
        '404':
          description: Message not found

    put:
      tags: [Slack]
      summary: Update message
      parameters:
        - name: channelId
          in: path
          required: true
        - name: ts
          in: path
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                blocks:
                  type: array
      responses:
        '200':
          description: Message updated

    delete:
      tags: [Slack]
      summary: Delete message
      parameters:
        - name: channelId
          in: path
          required: true
        - name: ts
          in: path
          required: true
      responses:
        '200':
          description: Message deleted
  /slack/messages:
    post:
      tags: [Slack]
      summary: Send message (alternative endpoint)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
                - channelId
              properties:
                text:
                  type: string
                  minLength: 1
                channelId:
                  type: string
                sessionId:
                  type: string
      responses:
        '201':
          description: Message sent

  # Calendar Endpoints
...